# loms/api/openapiv2/loms.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LOMS (Logistics and Order Management System) service",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Orders",
      "description": "Service for managing orders"
    },
    {
      "name": "Stock",
      "description": "Service for managing stock"
    }
  ],
  "host": "localhost:8082",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/order/cancel": {
      "post": {
        "operationId": "Orders_OrderCancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ordersOrderCancelRequest"
            }
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/order/create": {
      "post": {
        "operationId": "Orders_OrderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ordersOrderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ordersOrderCreateRequest"
            }
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/order/info": {
      "post": {
        "operationId": "Orders_OrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ordersOrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ordersOrderInfoRequest"
            }
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/order/pay": {
      "post": {
        "operationId": "Orders_OrderPay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ordersOrderPayRequest"
            }
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/v1/stock/info": {
      "post": {
        "operationId": "Stock_StocksInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/ordersStocksInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ordersStocksInfoRequest"
            }
          }
        ],
        "tags": [
          "Stock"
        ]
      }
    }
  },
  "definitions": {
    "ordersOrderCancelRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "ordersOrderCreateRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ordersOrderItem"
          }
        }
      }
    },
    "ordersOrderCreateResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "ordersOrderInfoRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "ordersOrderInfoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/ordersOrderStatus"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ordersOrderItem"
          }
        }
      }
    },
    "ordersOrderItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "ordersOrderPayRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "ordersOrderStatus": {
      "type": "string",
      "enum": [
        "NEW",
        "AWAITING_PAYMENT",
        "FAILED",
        "PAYED",
        "CANCELLED"
      ],
      "default": "NEW"
    },
    "ordersStocksInfoRequest": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "ordersStocksInfoResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}